Safari-611.1.10/LayoutTests/media/media-play-promise-reject-play-notsupported-expected.txt


RUN(mediaElement.src = findMediaFile("video", "content/invalid"))
EVENT(error)
RUN(mediaElement.play().then(failTest).catch(promiseRejected))
Promise rejected. OK
EXPECTED (error.name == 'NotSupportedError') OK
END OF TEST
